Added resource types. Tracing modules during intitialization registers resources types they are going to track. And if more than one resource type is going to be tracked, then the resource type id is also written in function call records. Added shmseg module for shared memory segment tracking.

+++ src/modules/sp_context_impl.c
@@ -47,7 +47,7 @@ char sp_context_registry[SP_CONTEXT_REGISTRY_SIZE][SP_CONTEXT_NAME_SIZE];
 unsigned int sp_context_create(const char* name)
 {
 	int index = sync_fetch_and_add(&sp_context_index, 1);
-	if (index >= SP_CONTEXT_REGISTRY_SIZE) {
+	if (index >= (int)SP_CONTEXT_REGISTRY_SIZE) {
 		--sp_context_index;
 		return 0;
 	}
